+++ src/vue/pages/parachains/tabs/crowdloan/CrowdloanCompletedList.vue
@@ -1,0 +1,49 @@
+<template>
+  <div class="crowdloan-completed-list">
+    <h1 class="crowdloan-completed-list__header">
+      {{ $t('parachains-page.crowdloan-completed-list.completed-header') }}
+    </h1>
+    <template v-if="true">
+      <template v-if="false">
+        <div class="crowdloan-completed-list__body"/>
+      </template>
+      <template v-else>
+        <no-data-message
+          class="crowdloan-completed-list__no-data"
+          :message="
+            $t('parachains-page.crowdloan-completed-list.no-data-message')
+          "
+          is-row-block
+        />
+      </template>
+    </template>
+    <template v-else>
+      <skeleton-loader />
+    </template>
+  </div>
+</template>
+
+<script>
+import SkeletonLoader from '@/vue/common/SkeletonLoader'
+import NoDataMessage from '@/vue/common/NoDataMessage'
+
+export default {
+  name: 'crowdloan-completed-list',
+
+  components: {
+    SkeletonLoader,
+    NoDataMessage,
+  },
+}
+</script>
+
+<style lang="scss" scoped>
+@import '~@scss/mixins';
+@import '~@scss/variables';
+
+.crowdloan-completed-list__header {
+  padding: 0 1.6rem;
+  margin-bottom: 2rem;
+}
+
+</style>